new chages

+++ controllers/addShopProductController.js
@@ -1,7 +1,9 @@
-import express from "express";
 import ApiGenericResponse from "../middleware/ApiGenericResponse.js";
 import GENERIC_RESPONSE_MESSAGES from "../enums/genericResponseEnums.js";
 import ShoProductpModel from "../models/shopProduct.js";
+import { getHeaders } from "../middleware/auth.js";
+import { getShopWithUserId } from "./shopCRUDController.js";
+
 const addShopProductController = async (req, res, next) => {
   const { product_name, product_description, price, photos, quantity, shop_id, unit } = req.body;
   if (!product_name) return next(ApiGenericResponse.badRequest(GENERIC_RESPONSE_MESSAGES.PRODUCT_NAME_REQUIRED, undefined, false));
@@ -9,6 +11,11 @@ const addShopProductController = async (req, res, next) => {
   if (!price) return next(ApiGenericResponse.badRequest(GENERIC_RESPONSE_MESSAGES.PRODUCT_PRICE_REQUIRED, undefined, false));
   if (!quantity) return next(ApiGenericResponse.badRequest(GENERIC_RESPONSE_MESSAGES.PRODUCT_QUANTITY_REQUIRED, undefined, false));
   try {
+    const { loginuserid } = getHeaders(req);
+    const isRightAccess = await validatingUserCanReadWriteProduct(loginuserid, shop_id);
+    if (isRightAccess) {
+      return next(ApiGenericResponse.successServerCode(GENERIC_RESPONSE_MESSAGES.NO_PERMISSION, undefined, false));
+    }
     const product = await ShoProductpModel({ product_name, product_description, price, photos, quantity, shop_id, last_updated: new Date(), created_at: new Date(), unit });
     if (product) {
       const registeredProduct = await product.save();
@@ -45,6 +52,11 @@ const getProductsWithShopId = async (shop_id) => {
 const updateProduct = async (req, res, next) => {
   const { product_name, product_description, price, quantity, unit } = req.body;
   try {
+    const { loginuserid } = getHeaders(req);
+    const isRightAccess = await validatingUserCanReadWriteProduct(loginuserid, req.body._id);
+    if (isRightAccess) {
+      return next(ApiGenericResponse.successServerCode(GENERIC_RESPONSE_MESSAGES.NO_PERMISSION, undefined, false));
+    }
     const updatedProduct = await ShoProductpModel.update({ _id: req.body._id }, { $set: { product_name, product_description, price, quantity, last_updated: new Date(), unit } });
     if (updatedProduct) {
       return next(ApiGenericResponse.successServerCode(GENERIC_RESPONSE_MESSAGES.SUCCESS, updatedProduct, true));
@@ -58,8 +70,13 @@ const updateProduct = async (req, res, next) => {
 };
 
 const deleteProductById = async (req, res, next) => {
-  const { id } = req.params;
   try {
+    const { loginuserid } = getHeaders(req);
+    const { id } = req.params;
+    const isRightAccess = await validatingUserCanReadWriteProduct(loginuserid, id);
+    if (isRightAccess) {
+      return next(ApiGenericResponse.successServerCode(GENERIC_RESPONSE_MESSAGES.NO_PERMISSION, undefined, false));
+    }
     const deletedProduct = await ShoProductpModel.deleteOne({ _id: id });
     if (!deletedProduct) {
       return next(ApiGenericResponse.internalServerError(GENERIC_RESPONSE_MESSAGES.INTERNAM_SERVER_ERROR, undefined, false));
@@ -67,8 +84,26 @@ const deleteProductById = async (req, res, next) => {
       return next(ApiGenericResponse.successServerCode(GENERIC_RESPONSE_MESSAGES.DELETE_SHOP_SUCCESS, undefined, true));
     }
   } catch (er) {
-    return next(ApiGenericResponse.internalServerError(GENERIC_RESPONSE_MESSAGES.INTERNAM_SERVER_ERROR));
+    console.error(err);
+    return next(ApiGenericResponse.internalServerError(GENERIC_RESPONSE_MESSAGES.INTERNAM_SERVER_ERROR, undefined, false));
   }
 };
-
+const getProductById = async (productId) => {
+  try {
+    return await ShoProductpModel.findOne({ _id: productId });
+  } catch (err) {
+    console.error(err);
+  }
+};
+const validatingUserCanReadWriteProduct = async (userId, productId) => {
+  const product = await getProductById(productId);
+  if (product) {
+    const shop = await getShopWithUserId(product.shop_id);
+    if (shop && shop.shop_owner_user_id !== userId) {
+      return true;
+    } else {
+      return false;
+    }
+  }
+};
 export { addShopProductController, getProductsWithShopId, updateProduct, deleteProductById };
