+++ controllers/shopCRUDController.js
@@ -4,6 +4,8 @@ import GENERIC_RESPONSE_MESSAGES from "../enums/genericResponseEnums.js";
 import { VidishaBazaarUser } from "../models/userModel.js";
 import { getProductsWithShopId } from "./addShopProductController.js";
 import GenericShopModel from "../services/GenericShopModel.js";
+import { getHeaders } from "../middleware/auth.js";
+
 const creatingShop = async (req, res, next) => {
   const { shop_name, shop_gst_number, shop_owner_user_id, shop_address, shop_city, shop_pincode, shop_mobile, shop_category_id, shop_image, is_shop_Physically_available, shop_id, description, shop_tags, opening_time, closing_time, days, shopSearchString } = req.body;
   if (!shop_name) return next(ApiGenericResponse.badRequest(GENERIC_RESPONSE_MESSAGES.SHOP_NAME_REQUIRED, undefined, false));
@@ -12,7 +14,7 @@ const creatingShop = async (req, res, next) => {
   if (!shop_category_id) return next(ApiGenericResponse.badRequest(GENERIC_RESPONSE_MESSAGES.SHOP_CATEGORY_ID_REQUIRED, undefined, false));
   if (!shop_address) return next(ApiGenericResponse.badRequest(GENERIC_RESPONSE_MESSAGES.SHOP_ADDRESS_REQUIRED, undefined, false));
   try {
-    const shop = await ShopModel({ shop_name, shop_gst_number: shop_gst_number, shop_owner_user_id, shop_address, shop_city, shop_pincode, shop_mobile, shop_category_id, shop_image, is_shop_Physically_available, last_updated: new Date(), created_at: new Date(), is_shop_varified: false, is_shop_active: false, is_shop_Physically_available: true, shop_id, description, shop_tags: shop_tags, opening_time, closing_time, days, shopSearchString });
+    const shop = await ShopModel({ shop_name, shop_gst_number: shop_gst_number, shop_owner_user_id, shop_address, shop_city, shop_pincode, shop_mobile, shop_category_id, shop_image, is_shop_Physically_available, last_updated: new Date(), created_at: new Date(), is_shop_varified: false, is_shop_active: false, is_shop_Physically_available: true, shop_id, description, shop_tags: shop_tags, opening_time, closing_time, days, shopSearchString, otp: Math.floor(100000 + Math.random() * 900000) });
     if (shop) {
       const registeredShop = await shop.save();
       if (registeredShop) {
@@ -43,19 +45,19 @@ const getAllShops = async (req, res, next) => {
         .sort({ _id: 1 })
         .skip(pageNumber > 0 ? (pageNumber - 1) * nPerPage : 0)
         .limit(nPerPage);
-      shopsCount = shops.length;
+      shopsCount = await ShopModel.count({ shop_category_id: subCategoryId, $text: { $search: searchString } });
     } else if (subCategoryId) {
       shops = shops = await ShopModel.find({ shop_category_id: subCategoryId })
         .sort({ _id: 1 })
         .skip(pageNumber > 0 ? (pageNumber - 1) * nPerPage : 0)
         .limit(nPerPage);
-      shopsCount = shops.length;
+      shopsCount = await ShopModel.count({ shop_category_id: subCategoryId });
     } else if (searchString) {
       shops = await ShopModel.find({ $text: { $search: `${searchString ? searchString : ""} ${subCategoryId ? subCategoryId : ""}` } })
         .sort({ _id: 1 })
         .skip(pageNumber > 0 ? (pageNumber - 1) * nPerPage : 0)
         .limit(nPerPage);
-      shopsCount = shops.length;
+      shopsCount = await ShopModel.count({ $text: { $search: `${searchString ? searchString : ""} ${subCategoryId ? subCategoryId : ""}` } });
     } else {
       shops = await ShopModel.find({})
         .skip(pageNumber > 0 ? (pageNumber - 1) * nPerPage : 0)
@@ -126,8 +128,13 @@ const getShopsWithShopIds = async (shopIds) => {
 };
 
 const deleteShopById = async (req, res, next) => {
-  const { id } = req.params;
   try {
+    const { loginuserid } = getHeaders(req);
+    const { id } = req.params;
+    const isRightUser = await validatingUserCanDeleteShop(loginuserid, id);
+    if (isRightUser) {
+      return next(ApiGenericResponse.successServerCode(GENERIC_RESPONSE_MESSAGES.NO_PERMISSION, undefined, false));
+    }
     const deletedShop = await ShopModel.deleteOne({ _id: id });
     if (!deletedShop) {
       return next(ApiGenericResponse.internalServerError(GENERIC_RESPONSE_MESSAGES.INTERNAM_SERVER_ERROR, undefined, false));
@@ -142,7 +149,7 @@ const deleteShopById = async (req, res, next) => {
 const updateShop = async (req, res, next) => {
   const { shop_name, shop_address, shop_mobile, shop_category_id, shop_image, description, shop_id, shop_tags, opening_time, closing_time, days, shopSearchString } = req.body;
   try {
-    const updatedShop = await ShopModel.update({ _id: req.body._id }, { $set: { shop_name, shop_address, shop_mobile, shop_category_id, shop_image, description, shop_id, shop_tags, last_updated: new Date(), opening_time, closing_time, days, shopSearchString } });
+    const updatedShop = await ShopModel.findOneAndUpdate({ _id: req.body._id }, { $set: { shop_name, shop_address, shop_mobile, shop_category_id, shop_image, description, shop_id, shop_tags, last_updated: new Date(), opening_time, closing_time, days, shopSearchString } });
     if (updatedShop) {
       return next(ApiGenericResponse.successServerCode(GENERIC_RESPONSE_MESSAGES.SUCCESS, updatedShop, true));
     } else {
@@ -158,4 +165,44 @@ const getRandomShops = async (limit) => {
   return await ShopModel.find().limit(limit).sort({ _id: 1 });
 };
 
-export { creatingShop, getAllShops, getShopsWithName, getShopWithId, getShospWithUserId, deleteShopById, updateShop, getShopsIdsByUserId, getShopsWithShopIds, getRandomShops };
+const getShopWithUserId = async (id) => {
+  try {
+    return await ShopModel.findOne({ _id: id });
+  } catch (err) {
+    console.error(err);
+  }
+};
+
+const validatingUserCanDeleteShop = async (userId, shopId) => {
+  const shop = await getShopWithUserId(shopId);
+  if (shop && userId !== shop.shop_owner_user_id) {
+    return true;
+  } else {
+    return false;
+  }
+};
+
+const otpVarificationForShop = async (req, res, next) => {
+  try {
+    const { otp, shopId } = req.body;
+    const { loginuserid } = getHeaders(req);
+    if (!otp) return next(ApiGenericResponse.successServerCode(GENERIC_RESPONSE_MESSAGES.OTP_REQUIRED, undefined, false));
+    if (!shopId) return next(ApiGenericResponse.successServerCode(GENERIC_RESPONSE_MESSAGES.SHOP_ID_REQUIRED, undefined, false));
+    const shop = await ShopModel.findOne({ _id: shopId });
+    if (shop && loginuserid && shop.shop_owner_user_id === loginuserid) {
+      if (shop.otp === otp) {
+        await ShopModel.update({ _id: shopId }, { $set: { is_shop_varified: true } });
+        return next(ApiGenericResponse.successServerCode(GENERIC_RESPONSE_MESSAGES.SHOP_SUCCESS_VARIFIED, undefined, true));
+      } else {
+        return next(ApiGenericResponse.successServerCode(GENERIC_RESPONSE_MESSAGES.OTP_NOT_MATCHING, undefined, false));
+      }
+    } else {
+      return next(ApiGenericResponse.successServerCode(GENERIC_RESPONSE_MESSAGES.USER_NOT_FOUND, undefined, false));
+    }
+  } catch (err) {
+    console.error("otp---", err);
+    return next(ApiGenericResponse.successServerCode(GENERIC_RESPONSE_MESSAGES.INTERNAM_SERVER_ERROR, undefined, false));
+  }
+};
+
+export { creatingShop, getAllShops, getShopsWithName, getShopWithId, getShospWithUserId, deleteShopById, updateShop, getShopsIdsByUserId, getShopsWithShopIds, getRandomShops, getShopWithUserId, otpVarificationForShop };
